setting of location using google maps

+++ app/src/main/java/com/example/kwako/models/House.java
@@ -92,7 +92,7 @@ public class House implements Parcelable {
         return lon;
     }
 
-    public void setLon(long lon) {
+    public void setLon(double lon) {
         this.lon = lon;
     }
